перенес стили ком-ов LeftBlock, RightBlock, UserInfoHeader, Wall, UserAbout, WallCreateArticle
