+++ src/common/leftBlock/index.jsx
@@ -1,9 +1,19 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-
-import { LeftBlockContainer } from '../styledComponents';
+import styled from 'styled-components';
 import Sidebar from '../sidebar/Sidebar';
 import ModalChat from '../modal-chat';
+
+const LeftBlockContainer = styled.div`
+  margin-top: 109px;
+  background: #111;
+  min-height: 100%;
+  flex-basis: 40%;
+  padding: 23px;
+  color: white;
+  display: flex;
+  flex-direction: column;
+`;
 
 const LeftBlock = ({ messages }) => (
   <LeftBlockContainer>
