+++ src/components/Main/UserAbout/UserAbout.styled.ts
@@ -1,0 +1,35 @@
+import styled from 'styled-components';
+
+export const WallInfoUserAbout = styled.div`
+  width: 531px;
+  border-top: 1px solid #515151;
+`;
+
+export const InfoHeaderText = styled.div`
+  margin-top: 47px;
+  font-size: 16px;
+  line-height: 20px;
+  color: #000;
+`;
+
+export const InfoHeaderTextBlock = styled.div`
+  margin-top: 47px;
+  display: flex;
+  flex-wrap: wrap;
+`;
+
+export const InfoHeaderListItemLeft = styled.div`
+  margin-bottom: 23px;
+  font-size: 16px;
+  line-height: 20px;
+  color: #515151;
+  flex-basis: 45%;
+`;
+
+export const InfoHeaderListItemRight = styled.div`
+  margin-bottom: 23px;
+  font-size: 16px;
+  line-height: 20px;
+  color: #000;
+  flex-basis: 45%;
+`;
