+++ src/components/Main/UserInfoHeader/UserInfoHeader.tsx
@@ -2,21 +2,21 @@
 import React, { useCallback } from 'react';
 import { connect, ConnectedProps } from 'react-redux';
 import addPhotoIcon from './img/add photo.svg';
+import ModalLinkInput from '../../../common/modalLinkInput';
+import { updateAvatar } from '../../../redux-toolkit/currentUserSlice';
+import { IUser } from '../../../types/user';
 import {
   UserInfoHeaderContainer,
   UserInfoAvatar,
   UserInfoNameBlock,
-  Avatar,
+  UserAvatar,
   AddPhotoBlock,
   AddPhotoIcon,
   UserName,
   UserProfession,
   UserOnlineStatus,
   UserOnlineIcon,
-} from '../../../common/styledComponents';
-import ModalLinkInput from '../../../common/modalLinkInput';
-import { updateAvatar } from '../../../redux-toolkit/currentUserSlice';
-import { IUser } from '../../../types/user';
+} from './UserInfoHeader.styled';
 
 const renderAddPhotoBlock = (setIsOpen: (newValue: boolean) => void, isCurrentUser: boolean) => {
   if (!isCurrentUser) {
@@ -51,7 +51,7 @@ const UserInfoHeader: React.FC<Props> = ({ user, updateAvatar: _updateAvatar, is
   return (
     <UserInfoHeaderContainer>
       <UserInfoAvatar>
-        <Avatar small={avatar} large={avatar} alt={`Аватар ${firstName} ${lastName}`} />
+        <UserAvatar small={avatar} large={avatar} alt={`Аватар ${firstName} ${lastName}`} />
         {renderAddPhotoBlock(setIsOpen, isCurrentUser)}
         <ModalLinkInput
           onLinkSend={onLinkSend}
