+++ src/components/Main/Wall/Wall.tsx
@@ -1,4 +1,10 @@
 import React from 'react';
+import WallCreateArticle from '../WallCreateArticle';
+import FormStatus from './FormStatus';
+import BlockNotes from '../Articles/blockNotes/BlockNotes';
+import UserAbout from '../UserAbout';
+import { IUser } from '../../../types/user';
+import { ImageDto } from '../../../types/image';
 import {
   WallContainer,
   WallInfoBlock,
@@ -6,13 +12,7 @@ import {
   InfoUserPhoto,
   InfoHeaderText,
   WallInfoUserAbout,
-} from '../../../common/styledComponents';
-import WallCreateArticle from '../WallCreateArticle';
-import FormStatus from './FormStatus';
-import BlockNotes from '../Articles/blockNotes/BlockNotes';
-import UserAbout from '../UserAbout';
-import { IUser } from '../../../types/user';
-import { ImageDto } from '../../../types/image';
+} from './Wall.styles';
 
 const renderPhotoBlock = (photos: ImageDto[] | null) => {
   if (!photos) {
